+++ src/components/currencies/currenciesList/paginationBar/PaginationBar.module.scss
@@ -1,0 +1,10 @@
+.pagination {
+  margin-top: 1.5rem;
+  justify-content: center;
+
+  button {
+    background-color: #222642;
+    border: none;
+    color: var(--color-white);
+  }
+}
